requirements.txt
aiogram
python-dotenv
Pillow
gspread
google-auth